Add files via upload
The credit department wants to simplify the current loan approval decision making process for the home equity line of credit. The current process is manual and prone to human error and biases. In this project, the immediate objective is to build a classification model for predicting the probability of default for any given customer applications and provide the bank with recommendations for important features to consider when approving a loan. The model will adopt guidelines of the Equal Credit Opportunity Act to establish an empirically derived and statistically sound credit scoring framework. The model will be based on recent data from the existing underwriting process, including customer profile information and the loan performance of approved customers. It will be built using predictive modeling techniques, and it must be interpretable to provide customer justifications for adverse actions (i.e. application rejections). Other objectives would include gaining insights from the loan performance data to aid in the design of customer acquisition strategies and help the business to best utilize the model's predictions for decision-making.